requirements.txt
click==8.3.1
colorama==0.4.6
python-dotenv==1.2.2
